requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-database-url
Django
django-environ
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
drf-yasg
gunicorn
idna
inflection
oauthlib
packaging
psycopg2-binary
pycparser
PyJWT
python-decouple
python3-openid
pytz
PyYAML
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
uritemplate
urllib3
whitenoise